# api/swagger_server/models/purchase_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class PurchaseRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, payment_token: str=None):  # noqa: E501
        """PurchaseRequest - a model defined in Swagger

        :param payment_token: The payment_token of this PurchaseRequest.  # noqa: E501
        :type payment_token: str
        """
        self.swagger_types = {
            'payment_token': str
        }

        self.attribute_map = {
            'payment_token': 'payment_token'
        }
        self._payment_token = payment_token

    @classmethod
    def from_dict(cls, dikt) -> 'PurchaseRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PurchaseRequest of this PurchaseRequest.  # noqa: E501
        :rtype: PurchaseRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def payment_token(self) -> str:
        """Gets the payment_token of this PurchaseRequest.


        :return: The payment_token of this PurchaseRequest.
        :rtype: str
        """
        return self._payment_token

    @payment_token.setter
    def payment_token(self, payment_token: str):
        """Sets the payment_token of this PurchaseRequest.


        :param payment_token: The payment_token of this PurchaseRequest.
        :type payment_token: str
        """
        if payment_token is None:
            raise ValueError("Invalid value for `payment_token`, must not be `None`")  # noqa: E501

        self._payment_token = payment_token